reorganized components and css
