--- react_app/src/Accordion.js
@@ -1,53 +1,0 @@
-// import {Collapse, Card, CardBody, CardHeader} from 'reactstrap';
-// import "bootstrap/dist/css/bootstrap.min.css";
-
-// function utcConverter(utcTime) {
-//     var d = new Date(0);
-//     d.setUTCSeconds(utcTime)
-//     var d2 = d.getDay()
-//     return d2;
-// }
-
-
-// let days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
-// function dayConverter(dayNum) {
-//     console.log(days[dayNum])
-//     return days[dayNum]
-// }
-
-// function toTitleCase(str) {
-//     return str.replace(/\w\S*/g, function (txt) {
-//         return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-//     });
-// }
-
-// const Accordions = (props) => {
-//     <Accordion>
-//   <Accordion.Item eventKey="0">
-//     <Accordion.Header>Accordion Item #1</Accordion.Header>
-//     <Accordion.Body>
-//       Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-//       tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
-//       veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
-//       commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
-//       velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
-//       cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
-//       est laborum.
-//     </Accordion.Body>
-//   </Accordion.Item>
-//   <Accordion.Item eventKey="1">
-//     <Accordion.Header>Accordion Item #2</Accordion.Header>
-//     <Accordion.Body>
-//       Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-//       tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
-//       veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
-//       commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
-//       velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
-//       cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
-//       est laborum.
-//     </Accordion.Body>
-//   </Accordion.Item>
-// </Accordion>
-// }
-
-// export default Accordions;
