+++ react_app/src/Components/IndCard.js
@@ -1,10 +1,8 @@
 import React, { useState } from "react";
 import { Collapse, Card, CardBody, CardText, CardTitle, Col } from 'reactstrap';
-// import Chevron from "./Chevron.js";
 import "bootstrap/dist/css/bootstrap.min.css";
-import "../Cards/Cards.css";
-import TextForCards from "../../TextForCards";
-
+import "../CSS/Cards.css";
+import TextForCards from "./TextForCards";
 
 
 const IndCard = (props) => {
