+++ react_app/src/Components/TextForCards.js
@@ -1,10 +1,6 @@
 import React from "react";
+import {toTitleCase} from '../Utils/Converters.js';
 
-function toTitleCase(str) {
-    return str.replace(/\w\S*/g, function (txt) {
-        return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-    });
-}
 const TextForCards = (props) => {
     const { daily } = props
     return (
